Implemented a better alg for input processing.
*Brief algorithm overview.*
The flow is divided into 4 tasks:

1. Input validation

2. Valid input read

3. "mas-piano-validation" step

4. Show results

*Step 1: input validation
*
The general idea is to filter all arguments and keep only the JSON files
.
For every path in input check if it is a directory.

If it is, then evaluate the *direct* children.

If it is not, it means it is a file.
So check for the extension since only JSON files are allowed.

If the file is not a JSON file, warn the user (doesn't matter when).

*Breakpoint between Step 1 and the following steps*: if there are no JSON files to be read then tell the user and quit.

*Step 2: read files contents
*
This step should be async and parallel so that I/O can run concurrently.

*Step 3: validation!* 🎉

Nothing else to say here really.

*Step 4: Show the results!*

Show first the good news: so general validation result and per file results.

Then report collected warnings.